rewrite hans command to ts
